requirements.txt
requests
telegram
python-telegram-bot
python-firebase
Pillow